moved twitter helper function
